\name{cAsoc}
\alias{cAsoc}
\title{
Dialog for function findAssocs() from package (tm)
}
\description{
This function provide interface to function \code{\link[tm]{findAssocs}} from \pkg{tm} package
}
\usage{
cAsoc()
}
\details{
You can also acces this function using \code{Rcmdr} menu.\cr
From \code{Rcmdr} menu choose \emph{TextMining -> TermDocumentMatrix -> Find associations... }
}

\author{
Dzemil Lushija \email{dzemill@gmail.com}
}

\seealso{
See Also \code{\link[tm]{findAssocs}}
}
\keyword{ Text mining }

