\name{cChunks}
\alias{cChunks}
\title{
Dialog for function makeChunks() from package (tm)
}
\description{
This function provide interface to function \code{\link[tm]{makeChunks}} from \pkg{tm} package
}
\usage{
cChunks()
}
\details{
Thiss function allows you to split corpus into chunks.\cr
You can also acces this function using \code{Rcmdr} menu.\cr
From \code{Rcmdr} menu choose \emph{TextMining -> Edit corpus -> Make Chunks... }
}

\author{
Dzemil Lushija \email{dzemill@gmail.com}
}

\seealso{
See also \code{\link[tm]{makeChunks}}
}
\keyword{ Text mining  }

