\name{cDelete}
\alias{cDelete}

\title{
Dialog for function rm() from package (base)
}
\description{
This function provide interface to function \code{\link[base]{rm}} from \pkg{base} package
}
\usage{
cDelete()
}

\details{
From list of corpuses you can choose one to remove from workspace.\cr
You can also acces this function using \code{Rcmdr} menu.\cr
From \code{Rcmdr} menu choose \emph{TextMining -> Delete Corpus... }
}

\author{
Dzemil Lushija \email{dzemill@gmail.com}
}

\seealso{
See also \code{\link[base]{rm}}
}

\keyword{ package }

