\name{cDict}
\alias{cDict}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Dialog for function Dictionary() from package (tm)
}
\description{
This function provide interface to function \code{\link[tm]{Dictionary}} from \pkg{tm} package
}
\usage{
cDict()
}
%- maybe also 'usage' for other objects documented here.
\details{
With this command you can create Dictionary from list of words or from TermDocumentMatrix.\cr
You can also acces this function using \code{Rcmdr} menu.\cr
From \code{Rcmdr} menu choose \emph{TextMining -> TermDocumentMatrix -> Creating dictionary... }
}

\author{
Dzemil Lushija \email{dzemill@gmail.com}
}

\seealso{
See also \code{\link[tm]{Dictionary}}
}

\keyword{ Text mining  }

