\name{cDiss}
\alias{cDiss}
\title{
Dialog for function dissimilarity() from package (tm)
}
\description{
This function provide interface to function \code{\link[tm]{dissimilarity}} from \pkg{tm} package for use with TermDocumentMatrix
}
\usage{
cDiss()
}

\details{
Compute the dissimilarity between documents in a TermDocumentMatrix.\cr
You can also acces this function using \code{Rcmdr} menu.\cr
From \code{Rcmdr} menu choose \emph{TextMining -> Dissimilarity  -> For TermDocumentMatrix...}
}

\author{
Dzemil Lushija \email{dzemill@gmail.com}
}

\seealso{
See also \code{\link[tm]{dissimilarity}}
}
\keyword{ Text mining  }

