\name{cDissDoc}
\alias{cDissDoc}

\title{
Dialog for function dissimilarity() from package (tm)
}
\description{
This function provide interface to function \code{\link[tm]{dissimilarity}} from \pkg{tm} package
}
\usage{
cDissDoc()
}

\details{
Compute the dissimilarity between two explicit documents.\cr
You can also acces this function using \code{Rcmdr} menu.\cr
From \code{Rcmdr} menu choose \emph{TextMining -> Dissimilarity -> For text documents... }
}

\author{
Dzemil Lushija \email{dzemill@gmail.com}
}

\seealso{
See also \code{\link[tm]{dissimilarity}}
}

\keyword{ Text mining  }

