\name{cFreq}
\alias{cFreq}

\title{
Dialog for function findFreqTerms() from package (tm)
}
\description{
This function provide interface to function \code{\link[tm]{findFreqTerms}} from \pkg{tm} package
}
\usage{
cFreq()
}

\details{
This command allows you to find frequent terms in a TermDocumentMatrix.\cr
You can also acces this function using \code{Rcmdr} menu.\cr
From \code{Rcmdr} menu choose \emph{TextMining -> TermDocumentMatrix -> Find frequent terms...}
}

\author{
Dzemil Lushija \email{dzemill@gmail.com}
}

\seealso{
See also \code{\link[tm]{findFreqTerms}}
}

\keyword{ Text mining  }

