\name{cInspect}
\alias{cInspect}

\title{
Dialog for function inspect() from package (tm)
}
\description{
This function provide interface to function \code{\link[tm]{inspect}} from \pkg{tm} package
}
\usage{
cInspect()
}

\details{
Displays detailed information for a corpus.\cr
You can also acces this function using \code{Rcmdr} menu.\cr
From \code{Rcmdr} menu choose \emph{TextMining -> Edit corpus -> Inspect Corpus...}
}

\author{
Dzemil Lushija \email{dzemill@gmail.com}
}

\seealso{
See also \code{\link[tm]{inspect}}
}

\keyword{ Text mining  }

