\name{cInter}
\alias{cInter}

\title{
Dialog for function tm_intersect() from package (tm)
}
\description{
This function provide interface to function \code{\link[tm]{tm_intersect}} from \pkg{tm} package
}
\usage{
cInter()
}

\details{
Perform intersection on text documents and words.\cr
You can also acces this function using \code{Rcmdr} menu.\cr
From \code{Rcmdr} menu choose \emph{TextMining -> Filters  -> Intersection...}
}

\author{
Dzemil Lushija \email{dzemill@gmail.com}
}

\seealso{
See also \code{\link[tm]{tm_intersect}}
}

\keyword{ Text mining  }

