\name{cmData}
\alias{cmData}

\title{
Dialog for function meta() from package (tm)
}
\description{
This function provide interface to function \code{\link[tm]{meta}} from \pkg{tm} package
}
\usage{
cmData()
}

\details{
Display meta data of corpus.\cr
You can also acces this function using \code{Rcmdr} menu.\cr
From \code{Rcmdr} menu choose \emph{TextMining -> Edit corpus -> Meta Data...}
}

\author{
Dzemil Lushija \email{dzemill@gmail.com}
}

\seealso{
See also \code{\link[tm]{meta}}
}

\keyword{ Text mining  }

