\name{cPlain}
\alias{cPlain}

\title{
Dialog for function as.PlainTextDocument() from package (tm)
}
\description{
This function provide interface to function \code{\link[tm]{as.PlainTextDocument}} from \pkg{tm} package
}
\usage{
cPlain()
}

\details{
Converts all documents in corpus to plane text documents using \code{\link[tm]{tm_map}}.\cr
You can also acces this function using \code{Rcmdr} menu.\cr
From \code{Rcmdr} menu choose \emph{TextMining -> Transformations -> To plain text documents...}
}

\author{
Dzemil Lushija \email{dzemill@gmail.com}
}

\seealso{
See also \code{\link[tm]{as.PlainTextDocument}}
}

\keyword{ Text mining  }

