\name{cPlot}
\alias{cPlot}

\title{
Dialog for function plot() from package (tm)
}
\description{
This function provide interface to function \code{\link[tm]{plot}} from \pkg{tm} package
}
\usage{
cPlot()
}

\details{
Visualize correlations between terms of a term-document matrix.\cr
You can also acces this function using \code{Rcmdr} menu.\cr
From \code{Rcmdr} menu choose \emph{TextMining -> Plot ... }
}

\author{
Dzemil Lushija \email{dzemill@gmail.com}
}

\seealso{
See also \code{\link[tm]{plot}}
}

\keyword{ Text mining  }

