\name{cRN}
\alias{cRN}

\title{
Dialog for function removeNumbers() from package (tm)
}
\description{
This function provide interface to function \code{\link[tm]{removeNumbers}} from \pkg{tm} package
}
\usage{
cRN()
}

\details{
Remove Numbers from a Text Documents in corpus using \code{\link[tm]{tm_map}}.\cr
You can also acces this function using \code{Rcmdr} menu.\cr
From \code{Rcmdr} menu choose \emph{TextMining -> Transformations -> Remove numbers... }
}

\author{
Dzemil Lushija \email{dzemill@gmail.com}
}

\seealso{
See also \code{\link[tm]{removeNumbers}}
}

\keyword{ Text mining  }

