\name{cRP}
\alias{cRP}

\title{
Dialog for function removePunctuation() from package (tm)
}
\description{
This function provide interface to function \code{\link[tm]{removePunctuation}} from \pkg{tm} package
}
\usage{
cRP()
}

\details{
Remove all punctuation marks from a text documents in corpus using \code{\link[tm]{tm_map}}.\cr
You can also acces this function using \code{Rcmdr} menu.\cr
From \code{Rcmdr} menu choose \emph{TextMining -> Transformations -> Remove punctuation...}
}

\author{
Dzemil Lushija \email{dzemill@gmail.com}
}

\seealso{
See also \code{\link[tm]{removePunctuation}}
}

\keyword{ Text mining  }

