\name{cSFT}
\alias{cSFT}

\title{
Dialog for function tm_filter() from package (tm)
}
\description{
This function provide interface to function \code{\link[tm]{tm_filter}} from \pkg{tm} package
}
\usage{
cSFT()
}

\details{
Interface to apply filter functions to corpus.\cr
You can also acces this function using \code{Rcmdr} menu.\cr
From \code{Rcmdr} menu choose \emph{TextMining -> Filters -> Search full text...}
}

\author{
Dzemil Lushija \email{dzemill@gmail.com}
}

\seealso{
See also \code{\link[tm]{tm_filter}}
}

\keyword{ Text mining  }

