\name{cSparse}
\alias{cSparse}

\title{
Dialog for function removeSparseTerms() from package (tm)
}
\description{
This function provide interface to function \code{\link[tm]{removeSparseTerms}} from \pkg{tm} package
}
\usage{
cSparse()
}

\details{
Remove sparse terms from a term-document matrix.\cr
You can also acces this function using \code{Rcmdr} menu.\cr
From \code{Rcmdr} menu choose \emph{TextMining -> TermDocumentMatrix -> Remove sparse terms...}
}

\author{
Dzemil Lushija \email{dzemill@gmail.com}
}

\seealso{
See also \code{\link[tm]{removeSparseTerms}}
}

\keyword{ Text mining  }

