\name{cStem}
\alias{cStem}

\title{
Dialog for function stemDocument() from package (tm)
}
\description{
This function provide interface to function \code{\link[tm]{stemDocument}} from \pkg{tm} package
}
\usage{
cStem()
}

\details{
Stem words in a all text documents in corpus using Porter's stemming algorithm.\cr
You can also acces this function using \code{Rcmdr} menu.\cr
From \code{Rcmdr} menu choose \emph{TextMining -> Transformations -> Stem documents...}
}

\author{
Dzemil Lushija \email{dzemill@gmail.com}
}

\seealso{
See also \code{\link[tm]{stemDocument}}
}

\keyword{ Text mining  }

