\name{cSummary}
\alias{cSummary}

\title{
Dialog for function summary() from package (base)
}
\description{
This function provide interface to function \code{\link[base]{summary}} from \pkg{base} package
}
\usage{
cSummary()
}

\details{
Display summary of Corpus.\cr
You can also acces this function using \code{Rcmdr} menu.\cr
From \code{Rcmdr} menu choose \emph{TextMining -> Edit corpus -> Corpus Summary...}
}

\author{
Dzemil Lushija \email{dzemill@gmail.com}
}

\seealso{
See also \code{\link[base]{summary}}
}

\keyword{ Text mining  }

