\name{cTolower}
\alias{cTolower}

\title{
Dialog for function tolower() from package (base)
}
\description{
This function provide interface to function \code{\link[base]{tolower}} from \pkg{base} package
}
\usage{
cTolower()
}

\details{
Translate characters from upper to lower case in all text documents in corpus.\cr
You can also acces this function using \code{Rcmdr} menu.\cr
From \code{Rcmdr} menu choose \emph{TextMining -> Transformations -> Convert to lowercase...}
}

\author{
Dzemil Lushija \email{dzemill@gmail.com}
}

\seealso{
See also \code{\link[base]{tolower}}
}

\keyword{ Text mining  }

