\name{cUtf8}
\alias{cUtf8}

\title{
Dialog for function enc2utf8() from package (base)
}
\description{
This function provide interface to function \code{\link[base]{enc2utf8}} from \pkg{base} package
}
\usage{
cUtf8()
}

\details{
Convert the encoding of a text documents in corpus to UTF-8.\cr
You can also acces this function using \code{Rcmdr} menu.\cr
From \code{Rcmdr} menu choose \emph{TextMining -> Transformations -> Convert to UTF-8...}
}

\author{
Dzemil Lushija \email{dzemill@gmail.com}
}

\seealso{
See also \code{\link[base]{enc2utf8}}
}

\keyword{ Text mining  }

