\name{cWhite}
\alias{cWhite}

\title{
Dialog for function stripWhitespace() from package (tm)
}
\description{
This function provide interface to function \code{\link[tm]{stripWhitespace}} from \pkg{tm} package
}
\usage{
cWhite()
}

\details{
Strip extra whitespace from all text documents in corpus. Multiple white space\cr characters are collapsed to a
single blank.\cr
You can also acces this function using \code{Rcmdr} menu.\cr
From \code{Rcmdr} menu choose \emph{TextMining -> Transformations -> Strip whitespaces...}
}

\author{
Dzemil Lushija \email{dzemill@gmail.com}
}

\seealso{
See also \code{\link[tm]{stripWhitespace}}
}

\keyword{ Text mining  }

