\name{cWords}
\alias{cWords}

\title{
Dialog for function removeWords() from package (tm)
}
\description{
This function provide interface to function \code{\link[tm]{removeWords}} from \pkg{tm} package
}
\usage{
cWords()
}

\details{
Remove a set of words from all text documents in corpus.\cr
You can also acces this function using \code{Rcmdr} menu.\cr
From \code{Rcmdr} menu choose \emph{TextMining -> Transformations -> Remove words...}
}

\author{
Dzemil Lushija \email{dzemill@gmail.com}
}

\seealso{
See also \code{\link[tm]{removeWords}}
}

\keyword{ Text mining  }

