\name{cWrite}
\alias{cWrite}

\title{
Dialog for function writeCorpus() from package (tm)
}
\description{
This function provide interface to function \code{\link[tm]{writeCorpus}} from \pkg{tm} package
}
\usage{
cWrite()
}

\details{
Write a plain text representation of a corpus to multiple files on disk \cr corresponding to the individual
documents in the corpus.\cr
You can also acces this function using \code{Rcmdr} menu.\cr
From \code{Rcmdr} menu choose \emph{TextMining -> Edit corpus -> Write Corpus...}
}

\author{
Dzemil Lushija \email{dzemill@gmail.com}
}

\seealso{
See also \code{\link[tm]{writeCorpus}}
}

\keyword{ Text mining  }

