\name{dirSource}
\alias{dirSource}

\title{
Dialog for function DirSource() from package (tm)
}
\description{
This function provide interface to function \code{\link[tm]{DirSource}} from \pkg{tm} package
}
\usage{
dirSource()
}

\details{
Create corpus from DirSource.\cr
You can also acces this function using \code{Rcmdr} menu.\cr
From \code{Rcmdr} menu choose \emph{TextMining -> Create Corpus -> Create DirSource...}
}

\author{
Dzemil Lushija \email{dzemill@gmail.com}
}

\seealso{
See also \code{\link[tm]{DirSource}}
}

\keyword{ Text mining  }

