\name{gSource}
\alias{gSource}

\title{
Dialog for function GmaneSource() from package (tm)
}
\description{
This function provide interface to function \code{\link[tm]{GmaneSource}} from \pkg{tm} package
}
\usage{
gSource()
}

\details{
Create corpus from GmaneSource.\cr
You can also acces this function using \code{Rcmdr} menu.\cr
From \code{Rcmdr} menu choose \emph{TextMining -> Create Corpus -> Create GmaneSource...}
}

\author{
Dzemil Lushija \email{dzemill@gmail.com}
}

\seealso{
See also \code{\link[tm]{GmaneSource}}
}

\keyword{ Text mining  }

