\name{RcmdrPlugin.TextMining-internal}
\title{Internal RcmdrPlugin.TextMining objects}
\alias{.Random.seed}
\alias{helpAbout}
\alias{manual}

\description{Internal RcmdrPlugin.TextMining objects and functions used in help menus}
\details{These are not to be called by the user.}
\usage{
  helpAbout()
  manual()
    }
\author{
Dzemil Lushija \email{dzemill@gmail.com}
}
\keyword{internal}