\name{rsSource}
\alias{rsSource}

\title{
Dialog for function ReutersSource() from package (tm)
}
\description{
This function provide interface to function \code{\link[tm]{ReutersSource}} from \pkg{tm} package
}
\usage{
rsSource()
}

\details{
Cretae corpus from Reuters Source.\cr
You can also acces this function using \code{Rcmdr} menu.\cr
From \code{Rcmdr} menu choose \emph{TextMining -> Create corpus -> Create ReutersSource...}
}

\author{
Dzemil Lushija \email{dzemill@gmail.com}
}

\seealso{
See also \code{\link[tm]{ReutersSource}}
}

\keyword{ Text mining  }

