\name{TDM}
\alias{TDM}

\title{
Dialog for function TermDocumentMatrix() from package (tm)
}
\description{
This function provide interface to function \code{\link[tm]{TermDocumentMatrix}} from \pkg{tm} package
}
\usage{
TDM()
}

\details{
Constructs a term-document matrix or a document-term matrix.\cr
You can also acces this function using \code{Rcmdr} menu.\cr
From \code{Rcmdr} menu choose \emph{TextMining -> TermDocumentMatrix -> Create TermDocumentMatrix...}
}

\author{
Dzemil Lushija \email{dzemill@gmail.com}
}

\seealso{
See also \code{\link[tm]{TermDocumentMatrix}}
}

\keyword{ Text mining  }

